+++ R/pheatmap-internal.R
@@ -5,8 +5,8 @@
 
 
 
-# Automatically handle the annotation data and colors.
-# Factors with a single level are automatically dropped.
+## Automatically handle the annotation data and colors.
+## Factors with a single level are automatically dropped.
 .pheatmapAnnotations <- function(
     object,
     blacklist = "sampleName",
@@ -18,11 +18,11 @@
         isHexColorFunction(legendColor, nullOK = TRUE)
     )
 
-    # Annotation columns -------------------------------------------------------
+    ## Annotation columns ------------------------------------------------------
     data <- colData(object)
     interestingGroups <- interestingGroups(object)
 
-    # pheatmap requires `NA` if annotations are empty.
+    ## pheatmap requires `NA` if annotations are empty.
     if (
         !hasDims(data) ||
         length(interestingGroups) == 0L ||
@@ -37,43 +37,42 @@
     )
     data <- data[, interestingGroups, drop = FALSE]
 
-    # Prepare the blacklist, always excluding sample names from labeling in
-    # the pheatmap annotation columns.
+    ## Prepare the blacklist, always excluding sample names from labeling in
+    ## the pheatmap annotation columns.
     blacklist <- unique(c("sampleName", blacklist))
 
     data <- data %>%
         as_tibble(rownames = "rowname") %>%
-        # Remove blacklisted columns (e.g. `sampleName`).
+        ## Remove blacklisted columns (e.g. `sampleName`).
         .[, setdiff(colnames(.), blacklist), drop = FALSE] %>%
-        # Ensure all strings are factors.
+        ## Ensure all strings are factors.
         mutate_if(is.character, as.factor) %>%
-        # Ensure unwanted numeric columns (e.g. sizeFactor) are dropped.
+        ## Ensure unwanted numeric columns (e.g. sizeFactor) are dropped.
         select_if(is.factor) %>%
         as.data.frame() %>%
         column_to_rownames("rowname")
 
-    # Drop any remaining factor columns that contain a single value.
-    # Note that we don't want to necessarily use `levels()` in place of
-    # `unique()` here, in case we have a situation where we're comparing a value
-    # against `NA`. Here this will a level of 1, even though we have 2 unique
-    # values.
+    ## Drop any remaining factor columns that contain a single value. Note that
+    ## we don't want to necessarily use `levels()` in place of `unique()` here,
+    ## in case we have a situation where we're comparing a value against `NA`.
+    ## Here this will a level of 1, even though we have 2 unique values.
     hasMultiple <- vapply(
         X = data,
         FUN = function(x) {
-            # This handles NA values better than using `levels()`.
+            ## This handles NA values better than using `levels()`.
             length(unique(x)) > 1L
         },
         FUN.VALUE = logical(1L)
     )
 
-    # Return empty if there are no useful factor columns.
+    ## Return empty if there are no useful factor columns.
     if (length(hasMultiple) == 0L) {
         return(.emptyPheatmapAnnotations)  # nocov
     } else {
         data <- data[, hasMultiple, drop = FALSE]
     }
 
-    # Colors -------------------------------------------------------------------
+    ## Colors ------------------------------------------------------------------
     if (
         is.data.frame(data) &&
         is.function(legendColor)
@@ -92,7 +91,7 @@
         colors <- NA
     }
 
-    # Return -------------------------------------------------------------------
+    ## Return ------------------------------------------------------------------
     list(
         annotationCol = data,
         annotationColors = colors
@@ -101,11 +100,11 @@
 
 
 
-# Sanitize formals into snake case and abort on duplicates.
-# Duplicates may arise if user is mixing and matching camel/snake case.
+## Sanitize formals into snake case and abort on duplicates. Duplicates may
+## arise if user is mixing and matching camel/snake case.
 .pheatmapArgs <- function(args) {
     assert(is.list(args), hasNames(args))
-    # Abort on snake case formatted formal args.
+    ## Abort on snake case formatted formal args.
     invalidNames <- grep("[._]", names(args), value = TRUE)
     if (length(invalidNames) > 0L) {
         stop(paste(
@@ -113,7 +112,7 @@
             toString(invalidNames)
         ))
     }
-    names(args) <- snake(names(args))
+    names(args) <- snakeCase(names(args))
     assert(
         isSubset(names(args), formalArgs(pheatmap)),
         hasNoDuplicates(names(args))
@@ -124,19 +123,19 @@
 
 
 
-# If `color = NULL`, use the pheatmap default palette
+## If `color = NULL`, use the pheatmap default palette.
 .pheatmapColorPalette <- function(color = NULL, n = 256L) {
     if (is.character(color)) {
-        # Hexadecimal color palette (e.g. RColorBrewer, viridis return).
+        ## Hexadecimal color palette (e.g. RColorBrewer, viridis return).
         assert(allAreHexColors(color))
         color
     } else if (is.function(color)) {
-        # Hexadecimal color function (e.g. viridis functions).
+        ## Hexadecimal color function (e.g. viridis functions).
         assert(isHexColorFunction(color))
         color(n = n)
     } else {
-        # pheatmap default palette.
-        # Note that `n` argument won't get evaluated here.
+        ## pheatmap default palette.
+        ## Note that `n` argument won't get evaluated here.
         eval(formals(pheatmap)[["color"]])
     }
 }