+++ R/plotCountsPerBroadClass-methods.R
@@ -6,7 +6,13 @@
 #' @param ... Additional arguments.
 #'
 #' @examples
-#' data(rse, sce, package = "acidtest")
+#' data(
+#'     RangedSummarizedExperiment,
+#'     SingleCellExperiment,
+#'     package = "acidtest"
+#' )
+#' rse <- RangedSummarizedExperiment
+#' sce <- SingleCellExperiment
 #'
 #' ## SummarizedExperiment ====
 #' plotCountsPerBroadClass(rse)
@@ -26,7 +32,8 @@ NULL
 
 
 
-plotCountsPerBroadClass.SummarizedExperiment <-  # nolint
+## Updated 2019-07-23.
+`plotCountsPerBroadClass,SummarizedExperiment` <-  # nolint
     function(
         object,
         assay = 1L,
@@ -49,12 +56,12 @@ plotCountsPerBroadClass.SummarizedExperiment <-  # nolint
             matchInterestingGroups(object, interestingGroups)
         interestingGroups <- interestingGroups(object)
 
-        # Get the count matrix.
+        ## Get the count matrix.
         assay <- assays(object)[[assay]]
-        # Ensure sparse matrix is coerced to dense.
+        ## Ensure sparse matrix is coerced to dense.
         assay <- as.matrix(assay)
 
-        # Log transform, if necessary.
+        ## Log transform, if necessary.
         if (trans == "log2") {
             assay <- log2(assay + 1L)
         } else if (trans == "log10") {
@@ -65,12 +72,12 @@ plotCountsPerBroadClass.SummarizedExperiment <-  # nolint
         }
 
         rowData <- rowData(object)
-        # Ensure Rle columns get decoded.
+        ## Ensure Rle columns get decoded.
         rowData <- decode(rowData)
         rownames(rowData) <- rownames(object)
 
         biotypeCol <- "broadClass"
-        # Warn and early return if the biotypes are not defined in rowData.
+        ## Warn and early return if the biotypes are not defined in rowData.
         if (!biotypeCol %in% colnames(rowData)) {
             warning(paste(
                 "rowData() does not contain biotypes defined in",
@@ -84,13 +91,13 @@ plotCountsPerBroadClass.SummarizedExperiment <-  # nolint
             select(!!sym(biotypeCol)) %>%
             group_by(!!sym(biotypeCol)) %>%
             summarise(n = n()) %>%
-            # Require at least 10 genes.
+            ## Require at least 10 genes.
             filter(!!sym("n") >= 10L) %>%
             arrange(desc(!!sym("n"))) %>%
             pull(!!sym(biotypeCol)) %>%
             as.character()
 
-        # Coerce the sample data to a tibble.
+        ## Coerce the sample data to a tibble.
         sampleData <- sampleData(object) %>%
             as_tibble(rownames = "sampleID") %>%
             mutate(!!sym("sampleID") := as.factor(!!sym("sampleID")))
@@ -103,7 +110,7 @@ plotCountsPerBroadClass.SummarizedExperiment <-  # nolint
                 -UQ(sym("rowname"))
             )
 
-        # SingleCellExperiment requires cell2sample mapping.
+        ## SingleCellExperiment requires cell2sample mapping.
         if (is(object, "SingleCellExperiment")) {
             c2s <- cell2sample(object, return = "tibble") %>%
                 rename(!!sym("colname") := !!sym("cellID"))
@@ -116,7 +123,7 @@ plotCountsPerBroadClass.SummarizedExperiment <-  # nolint
             data <- rename(data, !!sym("sampleID") := !!sym("colname"))
         }
 
-        # Prepare the minimal tibble required for plotting.
+        ## Prepare the minimal tibble required for plotting.
         data <- data %>%
             filter(!!sym("counts") > 0L) %>%
             left_join(
@@ -167,7 +174,7 @@ plotCountsPerBroadClass.SummarizedExperiment <-  # nolint
         p
     }
 
-formals(plotCountsPerBroadClass.SummarizedExperiment)[["fill"]] <-
+formals(`plotCountsPerBroadClass,SummarizedExperiment`)[["fill"]] <-
     formalsList[["fill.discrete"]]
 
 
@@ -177,5 +184,5 @@ formals(plotCountsPerBroadClass.SummarizedExperiment)[["fill"]] <-
 setMethod(
     f = "plotCountsPerBroadClass",
     signature = signature("SummarizedExperiment"),
-    definition = plotCountsPerBroadClass.SummarizedExperiment
+    definition = `plotCountsPerBroadClass,SummarizedExperiment`
 )