+++ NEWS.md
@@ -1,3 +1,17 @@
+## acidplots 0.2.7 (2019-07-23)
+
+### New functions
+
+- `plotCountsCorrelation` and `plotCountsCorrelationHeatmap`: New functions
+  optimized for plotting correlation between gene expression matrices.
+  Particularly useful for comparing pseudoaligned vs. aligned counts, which is
+  implemented in the new bcbioRNASeq package update.
+
+### Minor changes
+
+- Improved naming consistency of internal S4 functions.
+- Updated basejump dependencies.
+
 ## acidplots 0.2.6 (2019-07-18)
 
 ### Minor changes